[Fix] search 하위 폴더 위치 변경 및 SEO적용
search하위 url 수정으로 인해 폴더 구조 변경 및 SEO적용
