+++ front/components/search/resultList/ResultList.style.ts
