+++ front/components/search/resultList/ResultList.tsx
