--- front/pages/search/result/[pillName].tsx
@@ -1,7 +1,0 @@
-import FilteringSearch from "@searchComp/result/Result";
-
-function Filter() {
-  return <FilteringSearch />;
-}
-
-export default Filter;
